Unrelated: Moving Australia tests to their own test module
